+++ pkg/news.md
@@ -1,3 +1,4 @@
 ## Changes from Version 0.1.0 to 0.1.1
 
-* LassoShooting.fit -> handle of NAs in coefficient vector
+* LassoShooting.fit: handle of NAs in coefficient vector (Bug 1)
+* rlasso: removed bug in construction of model.matrix and change in predict-method (analog for rlassologit)